fix(lens): shortId build-id formatter placed on the rail consumer, not with the build identity
Moved the staleKey build-id formatter (shortId) from IdentityRail to kolu-common/surface as exported formatStaleKey beside PtyHostIdentitySchema, with IdentityRail importing it.

Agreed by the lowy ⇄ hickey lens debate (finding hickey-5, raised by hickey). Not pushed or merged.
